+++ cmd/login/flag.go
@@ -23,6 +23,9 @@ const (
 	flagWCInsecureNamespace = "insecure-namespace"
 
 	flagConnectorID = "connector-id"
+
+	flagProxy     = "proxy"
+	flagProxyPort = "proxy-port"
 )
 
 type flag struct {
@@ -40,6 +43,9 @@ type flag struct {
 	WCInsecureNamespace bool
 
 	ConnectorID string
+
+	Proxy     bool
+	ProxyPort int
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
@@ -57,6 +63,9 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().BoolVar(&f.WCInsecureNamespace, flagWCInsecureNamespace, false, fmt.Sprintf(`For client certificate creation. Allow using an insecure namespace for creating the client certificate. Requires --%s.`, flagWCName))
 
 	cmd.Flags().StringVar(&f.ConnectorID, flagConnectorID, "", "Dex connector to use for authentication. This allows to skip the selection page.")
+
+	cmd.Flags().BoolVar(&f.Proxy, flagProxy, false, "Enable socks proxy configuration for the cluster. Only Supported for Workload Cluster using clientcert auth mode")
+	cmd.Flags().IntVar(&f.ProxyPort, flagProxyPort, 9000, "Port for the socks proxy configuration for the cluster")
 
 	_ = cmd.Flags().MarkHidden(flagWCInsecureNamespace)
 	_ = cmd.Flags().MarkHidden("namespace")